Reworked manage provider to use 'username' instead of 'id' for the user's name
The original code ignores the use of the 'username' attribute.

This fix allows the use of 'username' which also bypasses some user name
limitations inherent to data bags.

For example, I wanted to have a user name with a period in it but the
data bag 'id' is limited to alphanumeric, -, _ characters.