[COOK-1583] Ability to set a password
When a 'password' entry exists in the users data bag, chef will set the user's
password.

Depenes on an Omnibus ruby, since it will ship with ruby-shadow. Not attempting
to make this backward compatible with non-omnibus rubies.

The password would not be saved to the data bag in clear text. It would be
hashed, which the user resource already accepts.  Hashing can be achieved with:
    `openssl passwd -1 "plaintextpassword"`