[COOK-3246] Allow customization of shell user env
It is sometimes desirable to allow for the customization of the shell
environment on a per-user basis.  Things like .vimrc or .my.cnf can
make life easier for shell users.

The philosophy here is that all files that customize a user environment will
appear in the users home directory.  The owner and group
of a file will always be the set to the user and their group, but file
permissions are exposed for control to simplify requirements such as those for
the `~/.ssh/authorized_keys` file.

See README.md for full details.