ensure default user group exists
The create user command will create a group by default. But if a user already exists, but their group does not, the group is never created and this provider will fail.

This change uses the group resource to ensure that the group exists. In normal situations this will be a NOOP but this catches the edge cases. If an explicit gid is specified for the user we can skip this because we assume that group is created/managed elsewhere.