Resolves Issue 118 fixing provider to work on Mac OS X
Added a .kitchen.vmware.yml that only works for Chef employees as
it uses the private images as available on Atlas.

Updated manage.rb to use the uid versus the username provided.
On Mac OS X, using the username for the owner field doesn't work
as expected. Additionally updated the group field to use gid but
only if it exists.